tests/wpt/tests/mediacapture-streams/GUM-impossible-constraint.https.html

<!doctype html>
<html>
<head>
<title>Trivial mandatory constraint in getUserMedia</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-NavigatorUserMedia-getUserMedia-void-MediaStreamConstraints-constraints-NavigatorUserMediaSuccessCallback-successCallback-NavigatorUserMediaErrorCallback-errorCallback">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#idl-def-NavigatorUserMediaError">
</head>
<body>
<p class="instructions">When prompted, accept to share your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that setting an impossible mandatory
constraint (width &gt;=1G) in getUserMedia works</p>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=permission-helper.js></script>
<script>
[
  // Note - integer conversion is weird for +inf and numbers > 2^32, so we
  // use a number less than 2^32 for testing.
  {width: {min:100000000}},
  {width: {max: 0}},
  {height: {max: 0}},
  {frameRate: {max: 0}},
  {width: {max: -1}},
  {height: {max: -1}},
  {frameRate: {max: -1}},
  {width: {min: 100, max: 10}},
  {height: {min: 100, max: 10}},
  {frameRate: {min: 100, max: 10}},
].forEach(constraints => promise_test(async t => {
  try {
    await navigator.mediaDevices.getUserMedia({video: constraints});
    assert_unreached('getUserMedia hould have rejected');
  } catch (err) {
    assert_equals(err.name, 'OverconstrainedError', "An impossible constraint triggers a OverconstrainedError");
    assert_equals(err.constraint, Object.keys(constraints)[0], "The name of the not satisfied constraint is given in error.constraint");
  }
}, `getUserMedia(${JSON.stringify(constraints)}) must fail with OverconstrainedError`));
</script>
</body>
</html>
